Switching the frontpage over to the 'retemplating' frontpage (the revised frontpage.html)
So basically, Paul's retemplating stuff for the frontpage was happening in index.html,
and the newer stuff I was doing after Paul went on leave (and after I made a mess of
the git repositories) was happening in frontpage.html, which was the same name as the
work pre-merge.

A lot of stuff needed to be re-added to make this work nice again, so re-added it all.